activemq-service/celery/requirements.txt
celery
stomp.py
kombu
sqlalchemy